insert IPINFO_TOKEN env at compile time
env var is inserted into binary at build time, while not being visible in the source code.
This means that the api token isn't fully protected, but that you'll have to be clever to find it.
For now, as the token is only valid for a free tier, this is okay.